+++ common/src/main/java/miyucomics/hexical/casting/operators/getters/OpGetItemStackData.kt
@@ -2,10 +2,11 @@ package miyucomics.hexical.casting.operators.getters
 
 import at.petrak.hexcasting.api.spell.ConstMediaAction
 import at.petrak.hexcasting.api.spell.casting.CastingContext
+import at.petrak.hexcasting.api.spell.getItemEntity
 import at.petrak.hexcasting.api.spell.iota.*
 import miyucomics.hexical.iota.IdentifierIota
-import miyucomics.hexical.iota.getItemStack
 import net.minecraft.enchantment.EnchantmentHelper
+import net.minecraft.item.ItemStack
 import net.minecraft.util.registry.Registry
 
 class OpGetItemStackData(private val mode: Int) : ConstMediaAction {
@@ -30,3 +31,5 @@ class OpGetItemStackData(private val mode: Int) : ConstMediaAction {
 		)
 	}
 }
+
+fun List<Iota>.getItemStack(idx: Int, argc: Int = 0): ItemStack = getItemEntity(idx, argc).stack