+++ forge/src/main/java/miyucomics/hexical/forge/HexicalForge.java
@@ -2,10 +2,7 @@ package miyucomics.hexical.forge;
 
 import dev.architectury.platform.forge.EventBuses;
 import miyucomics.hexical.Hexical;
-import miyucomics.hexical.HexicalClient;
-import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.eventbus.api.IEventBus;
-import net.minecraftforge.fml.DistExecutor;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;